+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub mod backend;
 pub mod cli;
 pub mod config;
+pub mod coordinator;
 pub mod core;
 mod entrypoint;
 pub mod notify;